Crate problem033

Problem 33 (Digit cancelling fractions)

Solution detail

There are four cases to consider, depending on which digit is cancelled from each side:

10a + b / 10a + d = b / d
10a + b / 10c + b = a / c
10a + b / 10b + d = a / d
10a + b / 10c + a = b / c

Expanding out each of these equations, we get to the following:

a(b - d) = 0
b(a - c) = 0
d(9a + b) = 10ab
b(9c + a) = 10ac

Of these, the first two represent trivial solutions, and the last two represent the same set of solutions, but with the two halves of the fraction switched. This means we can focus only on the third of these equations.

It is easy to brute-force solutions to this equation by iterating over a, b and checking if this gives an integer value for d.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the denominator of the product of all 2-digit cancelling fractions.